during service account recovery, remove invalid credentials
