+++ server/lib/src/idm/account.rs
@@ -371,16 +371,6 @@ impl Account {
         self.uuid == UUID_ANONYMOUS
     }
 
-    pub(crate) fn gen_generatedpassword_recover_mod(
-        &self,
-        cleartext: &str,
-        crypto_policy: &CryptoPolicy,
-    ) -> Result<ModifyList<ModifyInvalid>, OperationError> {
-        let ncred = Credential::new_generatedpassword_only(crypto_policy, cleartext)?;
-        let vcred = Value::new_credential("primary", ncred);
-        Ok(ModifyList::new_purge_and_set("primary_credential", vcred))
-    }
-
     pub(crate) fn gen_password_mod(
         &self,
         cleartext: &str,
